+++ examples/with-rax-mpa/src/pages/Home/index.tsx
@@ -5,6 +5,15 @@ import Logo from '@/components/Logo';
 
 import './index.css';
 
+class Rectangle {
+  height: any;
+  width: any;
+  constructor(height, width) {
+    this.height = height;
+    this.width = width;
+  }
+}
+
 export default function Home(props) {
   return (
     <View className="home">